minor entity id clean up
